+++ bioimageio/spec/_get_conda_env.py
@@ -187,32 +187,12 @@ def _get_default_onnx_env(*, opset_version: Optional[int]) -> BioimageioCondaEnv
 
 
 def _get_default_tf_env(tensorflow_version: Optional[Version]) -> BioimageioCondaEnv:
-    if tensorflow_version is None:
-        tensorflow_version = Version("1.15")
+    if tensorflow_version is None or tensorflow_version.major < 2:
+        tensorflow_version = Version("2.17")
 
-    # tensorflow 1 is not available on conda, so we need to inject this as a pip dependency
-    if tensorflow_version.major == 1:
-        tensorflow_version = max(
-            tensorflow_version, Version("1.13")
-        )  # tf <1.13 not available anymore
-        deps = (
-            "pip",
-            "python=3.7.*",  # tf 1.15 not available for py>=3.8
-            PipDeps(
-                pip=[
-                    "bioimageio.core",  # get bioimageio.core (and its dependencies) via pip as well to avoid conda/pip mix
-                    f"tensorflow =={tensorflow_version}",
-                    "protobuf <4.0",  # protobuf pin: tf 1 does not pin an upper limit for protobuf, but fails to load models saved with protobuf 3 when installing protobuf 4.
-                ]
-            ),
-        )
-        return BioimageioCondaEnv(
-            dependencies=list(deps),
-        )
-    else:
-        return BioimageioCondaEnv(
-            dependencies=["bioimageio.core", f"tensorflow =={tensorflow_version}"],
-        )
+    return BioimageioCondaEnv(
+        dependencies=["bioimageio.core", f"tensorflow =={tensorflow_version}"],
+    )
 
 
 def _get_env_from_deps(